Add files via upload
Basically it contains the basic multiplayer function, with some sort of turn concept(if your random user id is even you will be assigned red, else you will be assigned yellow, obviously this was done for testing purposes and needs to be worked upon). The rest is quite simple; just extend it to the full grid rather than just one box (which should be easy to implement since the communication between clients and server is based on box index), and just copy, paste and tweak a little the previously defined functions which make up the logic of the connect 4.